Setting max_byapass count to 16 instead of leaving at default value of 255 (#6581)
This setting affects how many times TE would cache the table data before
issuing read for the data when data is coming back-back.
With value of 255, when getting back-back traffic, if a program
at a stage is read/modifying, the updated data is not visible
to other programs running at diff stages. This very much affects
the RDMA/Rate Limitter/DCQCN usecases where data written in CB by
 one stage is needed to be available in otehr satges or in other pipe
lines.

Perhaps, Mike suggested this value to be 16 by default.
Hoping this does not affect any performance, but we will debug
more if issues seen after the commit.