Remove CMD dependencies on hostname and related fixes (#6580)
We have observed situations where the Venice cluster fails to form because
of mismatches between the node identifiers that user puts in the cluster
formation POST, hostnames and what hostnames resolve to.

For example:
- hostname is not set
- DNS name of the host is different from hostname
- user provides IP addresses but hostname resolves to a different address
- user provides an IP address but there are other addresses (possibly
  in the same subnet) configured on the interface

This typically results in Venice components not being able to connect or
authenticate each other.

This change tries to remove dependencies of CMD on hostname and DNS.
The general principle is that CMD tries to use exactly the node identifier
provided by the user in the initial POST. If it is an IP address,
CMD (and Venice components by extension) connect and authenticate based
on IP address. If it is a DNS name, they connect and authenticate based
on the name only. However, this is not always entirely possible.
For example, Etcd requires an IP:port pair to bind to and does not accept
a hostname. In that case, if user has provided a DNS name, we resolve
the DNS name (not hostname) and use the first returned address.

Other changes:
- Fixes https://github.com/pensando/sw/issues/6465
- introduces validation of user-provided node identifiers and
  rejects anything that is not a valid DNS name or IP address.
- fixes an issue by which CMD was not resolving host names
  before passing them to Etcd after a restart (startup.go path)
- use different auth directory for each run of Elastic integration tests