ionic: fix panic unbinding drivers
Ionic removes debugfs recursively before unregistering the net device, but the
rdma driver still has entries there.  That leaves rdma driver temporarily with
dangling pointers to its debugfs entries that were removed by the eth driver,
and then the panic is when rdma driver goes to remove its entries when it gets
the netdev unregister event.

Fix is to move debugfs_remove_recusive to after unregistering the netdev.  This
change moves it to the reverse order of operations that added debugfs in probe.