RDMA CM support (#7351)
* RDMA CM: Enable QP privileged mode

Kernel resources are privileged. RDMA CM only works on privileged
QP 1 created by the rdma kernel driver during driver load time.
This QP has to marked privileged to allow access to reserved lkey 0
that dma maps the entire kernel space.

Signed-off-by: Nitish Bhat <bnitish@pensando.io>

* RDMA CM: Set fence bit for eq interrupt assert dma cmd

EQ interrupts must be delivered to the host only after the eq has
been written to. Without the fence bit, there can be a race between
when the interrupt is delivered and the actual eq data arrives
at the host.

Signed-off-by: Nitish Bhat <bnitish@pensando.io>

* RDMA CM: Pass privileged bit info for UD QPs in Resp RX Pipeline

UD QP 1 is used exclusively by RDMA CM to communicate with peer
HCAs. This QP has to be marked privileged to load the reserved
lkey for incoming CM responses. Passing this info to the WQE stage
for incoming packets in resp_rx pipeline.

Signed-off-by: Nitish Bhat <bnitish@pensando.io>

* RDMA CM: CQ Pindex mirrors CQ Cindex in arming event

When an arming event occurs, instead of always resetting the
CQ cindex to 0, advance the pindex to mirror the cindex.
Proxy Pindex in the CQ is used to determine where to produce
new CQEs. This is more consistent with actual hw queue/ring
semantics.

Signed-off-by: Nitish Bhat <bnitish@pensando.io>

* RDMA CM: Move load_ah_process in req_tx to s3_t3

load_ah_process was loaded in t1 stage 3 in the req_tx pipeline.
It invalidates table 1 after it's done loading the ad size.
Parallel sge tables in the same stage use the same table 1 to load
sqlkey/rsvd lkey process in case there are more than 1 SGE in the
wqe. Based on timing, the load_ah process can end up invalidating
the table load command issued by parallel tables in the same stage.
Today, table 3 in stage 3 is not used. Also, this table is empty in
stage 4. So, using this table is safe for now. If tables are loaded
here in future, care needs to be taken to check RDMA CM is not broken.

Signed-off-by: Nitish Bhat <bnitish@pensando.io>

* RDMA CM: Move q_key updates to tblwr from memwr

Q-Key modify qp operation rewrites the pc offset in rqcb as the
offset is not word aligned. Move q_key updates to table writes
in individual stages.

Signed-off-by: Nitish Bhat <bnitish@pensando.io>

* RDMA CM: Add privileged q_key check for UD QPs

Privileged UD QPs can set a privileged Q_KEY (set high bit) from the
reserved range in the QPCB with a modify_qp operation. To use this
q_key in actual packets, the app has to just punch in the reserved
q_key (0x80000000) in the wqe q_key field. This is an indication
to the HCA to ignore the q_key in the wqe and instead use the
pre-programmed q_key in the QPCB. This is a privileged operation
and only kernel QPs can so this. Use of the reserved q_key (or
any q_key with the high bit set) by a regular QP results in
the phv being silently dropped.

Signed-off-by: Nitish Bhat <bnitish@pensando.io>

* RDMA CM: Do not imitialize timers for UD QPs

Recent change in the admin path introduced default values for retry cntr
and ack timeout. This and many other QP config params are optional or
not necessary for UD QPs. Setting ACK timeouts causes UD QPs to
re-transmit wqes after it has successfully completed them.

Signed-off-by: Nitish Bhat <bnitish@pensando.io>

* RDMA CM: Avoid LOCAL_WRITE access check for send, write operations

RDMA Send and Write operations do not mandate LOCAL_WRITE
permissions to be set on the memory region. Avoid this access
check (krping registers send bufs without WRITE permissions).

Signed-off-by: Nitish Bhat <bnitish@pensando.io>

* RDMA CM: Set max PMTU size for UD QPs

PMTU is not relevant for UD QPs. The spec says we must validate
the active PMTU for UD Qps and drop if payload length is greater
than the active PMTU. We set PMTU for UD QPs to max allowed in
IB spec, for now. Ideally we would also like to see active mtu
changes in the host and check for every wqe, including RC QPs.

Signed-off-by: Nitish Bhat <bnitish@pensando.io>