TLS for objectstore/minio (#7394)
Enable HTTPS on the REST interface exposed by Minio.
Note that this only allows the client to authenticate the server.
Minio does not support client-side certificates and client authentication
is performed using a shared secret. Right now this is hardcoded.
Eventually it will be stored in etcd and clients will pull it from there.