pnso_req: Add routine to route chain/batch poller based on poll-context flags
API layer returns chain or batch pointer as poll-context to
async/interrupt handler.  Upon an interrupt, when the handler calls-in
the API layer back with the poll-context, it needs to be determiined
whether the supplied-context was a chain or batch.  This commit adjusts
the existing code to rely on certain bits of batch/chain flags (i.e.
bi_flags/sc_flags) for the same.

Signed-off-by: Mohan Kumar <mohan@pensando.io>