Multi-pkt read response optimiziation - part 2 (#7704)
- Do not post completion for read_first as due to speculation, its not
possible to check msn in packet against msn in qstate in rrqwqe_process
to see if completion has been already posted for that msn. It cannot be
checked in sqcb1_write_back as well, as cmdeop has to be set either in
rrqptseg_process or in cqcb. If rrqptseg_process doesn't set then
cqcb_process has to set but in the event its a duplicate completion
neither of them will set.
- Check for duplicate completion is moved to sqcb1_write_back to make
it work for both speculation and non-speculation cases.
- read_req_adjust should be used only for the firt wqe upon start of
retransmission and not after that. Use read_req_adjust only if spec_cindex
is same as cindex (which will be the case at the start of
retransmission). Even if this phv gets dropped (due to rate limit or
some other reason), read_req_adjust is still maintained so that even
after spec reset read_request is transmitted only for the subset.
- When backtracking and spec reset happens at the same time,
backtracking will wait for spec_cindex to be same as cindex to make
sure sqcb state is not updated by any phvs in the pipeline. Because
spec_reset is set, cindex will never become same as spec_cindex and so
backtracking will never start. If spec_reset is set, revert spec_cindex
to cindex and start backtracking.