Stop corrupting non-RDMA AdminQ PHVs (#7096)
In RDMA Stage0 program, phv writes are done after ensuring
the PHV is targeted to one of the RDMA programs. For AdminQ,
just checking the queue type is not enough, as this type is
also shared with Eth and not exclusively reserved for RDMA
operations. AQ Qid 1 (onwards ?) belongs to RDMA, so check
this before issuing PHV writes..

Signed-off-by: Nitish Bhat <bnitish@pensando.io>