pnso_cp: Add shadow/restore support routines to tackle P4+ padding
P4 is updating the destination SGL, which is causing the sonic driver to
leak the SGLs.  This commit adds a hack to remember the list of dst SGLs
in src SGLs reserved field, and upon submitting/receving the update from
P4 layer, the SGLs are fixedup to prevent from leaking.  This fix has
been tested with 16K buffers locally, and more testing is pending.

Squashed:
	pnso_cp: Rename hack/unhack to shadow/restore routines
	pnso_cp: Restore SGLs only if shadowed

Signed-off-by: Mohan Kumar <mohan@pensando.io>