Backtrack Writeback (#7822)
SQCB0 and SQCB2 are updated simultaneously at the end of every backtrack
pass. SQCB0 writeback is protected with a busy bit but not SQCB2. As a
result in the next scheduling opportunity, a phv can read updated SQCB0
but old SQCB2, which means cindex is backtracked but not rexmit psn. This
can lead to retransmitting messaegs that were already sent as new messages
and there by posting completions with non-existing MSNs. Have a busy bit
for SQCB2 as well and drop phv if SQCB2 busy bit is still set and wait until
busy bit is unset to start the next pass of backtracking