linux ionic: use the common ionic_if file
Convert code to use the common ionic_if.h file, and remove a
couple of extraneous #includes.  Also tweak both the eth and
the rdma driver Makefiles to add the common directory to the
compiler's include list.  Also fix up the platform build to
get the common file in the right place for both the eth and
the rdma build.

Signed-off-by: Shannon Nelson <snelson@pensando.io>