ionic: fill out the stats dump struct
Following the notes from Neel, fill out the stats dump
struct to be used by the driver to read the network
statistics made available by the NIC.

Does this really need to be 4096 bytes, or can we shrink
it down to 1024 bytes?

Will there be counters for these netdev stats?
	collisions
	rx_length_errors
	rx_over_errors
	rx_crc_errors
	rx_frame_errors
	rx_fifo_errors
	tx_aborted_errors
	tx_carrier_errors
	tx_fifo_errors
	tx_heartbeat_errors
	tx_window_errors


Signed-off-by: Shannon Nelson <snelson@pensando.io>