ionic: fix up a bunch of checkpatch complaints
Running checkpatch.pl results in a bunch of trivial complaints that
needed to be cleaned up.  None of these are operational changes, simply
code formatting issues.

This patch doesn't clean all the complaints - there are a few that will
get cleaned later as a few more features get filled out.

Signed-off-by: Shannon Nelson <snelson@pensando.io>