ionic: free some memory
Several blocks of memory have been reserved using the devm_
mechanism that guarantees removal of the memory when the device
is removed.  However, these are chunks that may come and go
through the lifetime of the device, especially when we start
supporting multi-lif, and are not appropriate for depending
upon device removal - we'll be leaking memory.  We'll keep
with the devm_ mechanics for the safety it gives, but let's
still try to clean up after ourselves correctly.

Signed-off-by: Shannon Nelson <snelson@pensando.io>