linux ionic: add notifyq support
The ionic notifyq is how the NIC can tell the driver about
interesting events such as LinkChange or Reset.  It is set up
much like its other queues (adminq, txq, rxq), except that it
only uses the completion queue, in which the NIC writes the
events.  The command side of the qcq is only used to set up
the notifyq callback function pointer.

This patch adds the initial support for setting up the notifyq
and decoding events.  Later patches will add event reactions.

Signed-off-by: Shannon Nelson <snelson@pensando.io>