ionic: hack for rss key data passing to FW
Before the AdminQ was in place, dev_cmd was used to pass requests
and extra data was passed through the side_data space at
dev_cmd+0x800.  Evidently the FW is still expecting the RSS table
data to be there, rather than looking in the DMA shared space
indicated by the AdminQ command.

This is a HACK to get the data to the FW through the
dev_cmd interface rather than the documented AdminQ method.

Signed-off-by: Shannon Nelson <snelson@pensando.io>