Avoid propagating spurious NIC health updates to ApiServer (#8535)
Fixes https://pensando.atlassian.net/projects/VS/issues/VS-52

NAPLES agent sends periodic health updates to CMD, consisting of
a cluster.SmartNIC object populated with 1 "Condition" attribute
in Status field specifying NIC health. If the health status is
not changed, the only difference between this condition and
the condition on the corresponding ApiServer object is
LastTransitonTimeStamp.

CMD is responsible for changing the health status to "unknown" if
no updates are received from the agent within DeadTimeInterval.

With this PR, CMD maintains a local copy of NIC status and uses it
to keep track of agent updates. Updates are sent to ApiServer only if
there is an actual state change.

PR also includes:
  * an E2E test to verify that health is properly updated when NIC
    stops/restarts sending updates
  * logic to make DeadInterval and WatchInterval configurable to
    speed up tests in E2E
  * a generic "FilterUpdate" function to decide whether an update
    should be propagated to ApiServer or can be safely dropped
  * various related bug fixes