ionic: add reset lif to shutdown
Now that we have an interface defined to be used, ask the firmware
to reset the lif before we finish the cleanup work.  This will get
all NIC activity stopped in the lif stopped and in a stable state
for cleanup.

Note that we are adding the lif_reset command into the adminq_cmd
union in ionic_if.h, evidently this was forgotten in the patch that
added the interface.

We also fix a likely bug with trying to free the notifyblock after
we've already freed the netdev struct that houses the lif.

Signed-off-by: Shannon Nelson <snelson@pensando.io>