Fix race condition on veniceIntegSuite.TestTechSupportLabelBasedSelec… (#8646)
The test updates labels on nodes and then creates a new TechSupportRequest
to make sure that the new labels are honored. However, the controller receives
the node label update and techsupport request create events on two independent
watch channels and can process the techsupport request create before the node
label update, matching the old labels. The fix consists in waiting for the
node update to be processed by the controller before creating the new techsupport
request.