Fixes #8325 (#8683)
The test was failing because the expected event (e.g. evt.Count=15) was received in
different batches (e.g. batch1 -> evt.Count=12 and batch2 -> evt.Count=15) and
the alert engine was run on each batch. As a result the alert policy
will have a hit count 2 instead of 1 (expected from the test). We cannot control
this behavior as it depends on how fast the events are recorded from the
goroutine). So, the test case is changed to measure the accurate behavior.

It also fixes the below issue:
- Alert engine should not allow the alert to be created from the same event
  multiple times for the given policy.
  e.g. if p1 has a requirement of {count>10} i.e. create an alert when
  event.Count > 10. In this case if this possible that evt1 can initially come
  with count = 11 and later the same event can come with count=16 because of the
  deduplication mechanism we have in place. In this scenario, alert should be
  created only once on the first occurrence (evt1.Count=11).

Signed-off-by: Yuva Shankar <yuva29@users.noreply.github.com>