ionic: add support for macvlan hardware offload
Use our multi-lif capability to support macvlan hardware offload.
This allows us to dedicate a TxRx queue pair to one specific
mac address, which is represented in the Linux system as a
network pseudo-device of type macvlan and can be assigned to a
container, VM, or some other process separation gizmo.  These
macvlan pseudo-devices are upperdevs to the ionic lowerdev.

If the FW tells us that it supports multiple lifs, we will
advertise macvlan offload support to the network stack.
When a user does something like
    ip link add name mv1 link enp181s0 type macvlan
the stack will call into our .ndo_dfwd_add_station callback
with a new mac address to be filtered through a separate TxRx
queue pair.  The driver will create a new "slave" lif with a
single queue pair and set its filter to this new mac address.

The master lif's txqcqs[] and rxqcqs[] arrays are extended
beyond the number of queues that it actually uses to include
pointers to the slave lif qcqs.  When created, the qcqs in the
slave lifs will be linked into the next available slot in the
master lif's arrays in order for the master lif to find the
correct lif/queue when a Tx is requested.  From there, the
typical Tx and Rx processing happens seemlessly.

The initial model used in this driver was to init all the
lifs and give each its own netdev.  With this new usage model
where the additional lifs are the macvlan slaves, we don't
need to create all the lifs at once.  Also, only the master
lif will have a netdev, so the slaves need to create their
own lif struct.

Current known bug:
  The offloaded macvlans need to be removed/deleted before
  removing the ionic driver, else the system will hang
  waiting on a netdev reference count that never goes to 0.

Signed-off-by: Shannon Nelson <snelson@pensando.io>