ionic: macvlan addr change workaround
Given an offloaded macvlan, the kernel will crash with a NULL
pointer violation when someone tries to change the default mac
address with a command such as
	ifconfig mv1 hw ether 00:11:00:04:05:06

In macvlan_open() in Linux kernels older than v4.18, if the lower_dev
is advertising macvlan offload with the NETIF_F_HW_L2FW_DOFFLOAD bit
set in netdev->features, the driver's ndo_dfwd_add_station() is called.
If that succeeds, then macvlan_open() exits.  If no offload is advertised,
macvlan_open() does some device setup, including macvlan_hash_add(vlan)
which sets up vlan->hlist.  Since this doesn't get called for the
offload case, vlan->hlist is left uninitialized.  It is this uninitialized
bit that we stumble over later when we try to change the mac address.

The correct fix is a one line patch to macvlan.c to change the return
after the ndo_dfwd_add_station() call to "goto hash_add;"  However, we
don't have that luxury, so we have to fake it by hacking that code
here into our own driver.

PS-822

Signed-off-by: Shannon Nelson <snelson@pensando.io>