ionic: fix macvlan hang after link flap
PS-830 Ping stops working after flapping macvlan interface

After creating the macvlan interface, if someone stopped it
and restarted it with something like
	ifconfig mv1 down
	ifconfig mv1 up
the ping would not send out any packets.  This is because
the Tx queues were not restarted in the netdev interface
above the driver.  This is fixed by simply setting the
carrier up and waking the tx queues.

However, this exposed a new issue - the interface would hang,
and eventually we would start seeing soft lock reports from
the kernel.

This problem is that the macvlan driver's macvlan_stop() skipped
the call to macvlan_hash_del() when closing up an offloaded
macvlan interface, leaving the vlan->hlist still set in the
vlan hash list.  When the device was brough back up, it got
added again to the list with a circular pointer to itself.
Traversing that list for something like handling broadcast
packets now puts you into a loop.

By the look of bugs from other testing, it looks like it was
causing several odd things.  This is likely to address the
issues described in PS-849 and PS-850.

The fix is similar to the earlier macvlan_open() fix - we
copy the relevant code into our driver at the end of our
dfwd_del_station() and clean up the hlist reference.  This
one isn't quite as ugly as the previous hack.

Signed-off-by: Shannon Nelson <snelson@pensando.io>