Implement connection retries in certsproxy (#9369)
Have certsproxy make multiple attempts to connect to the Certificates API
before returning an error to the client. Before these changes, certsproxy
returned an API-level error to the client as soon as the first connection
attempt failed. This resulted in connections not being retried when
certsproxy was used.

Fixes sanity issue observed in http://jobd.pensando.io/logs/2053788