Do not re-init shm on a process restart (#9380)
1. Shm can be re-initialized in 2 cases:
	i) Process restarts - Delete the existing the shared memory and initialize
the new one. This way we need not have to deal with the old shared memory(might
have corrupted data/pointers).

	ii) Upgrades - During upgrade, each process will get a notifcation and
they can decide to delete underlying shared memory. Reader will still be
able to read the data using the existing handler and it can go away once
all the events are read. So, no code changes are needed to handle this
case.

2. Convert hal key to api object meta using the ntranslate interface.

Signed-off-by: Yuva Shankar <yuva29@users.noreply.github.com>