Dedup failed events (#9489)
So far, deduplication was not applied on failed events.
It was read from the file and sent as such to the exporters (venice,
syslog). But it is possible that there could be huge number of events
going out without dedup. So, applying dedup helps us reduce the volume
of events going out.

Also, this helps for debugging test failures.

Signed-off-by: Yuva Shankar <yuva29@users.noreply.github.com>