ionic: simplify device removal to fix unregister_netdev hang (#9515)
If the macvlan devices weren't deleted before the user did an rmmod,
the rmmod would hang the machine and these messages would be displayed:
    unregister_netdevice: waiting for enp181s0 to become free. Usage count = 1
This required a machine reset to recover: bad juju.

There is a bug in the kernel's netif_set_real_num_rx_queues() that
didn't remove all the objects that had a refcnt when the netdev was in
the NETDEV_UNREGISTERING state.  To get around this, we can simplify our
real_num_tx/rx_queues tracking to just tracking the Tx queues used by
the slaves and don't worry about the Rx queues.  At the same time, we
don't need to record the queue on which we received a packet for a
macvlan, as there won't be any useful RPS on the macvlan device anyway,
and that's where the real_num_rx_queues gets used in the stack.

Once that was fixed, we ran into a crash in our lif list handling in
ionic_unregister_netdev().  Since we know there really is only one netdev
that will be registered, just unregister it.  Any other lifs that exist
at that time will get taken care of as a side effect, such as macvlan
slaves that weren't removed.

PS-859

Signed-off-by: Shannon Nelson <snelson@pensando.io>