Add a null check to the pd_session_get before deferencing the PD data. (#9683)
When the session non-periodic timer get fired and if the session got
deleted between getting the session from hal handle hash table and
triggering a pd_session_get() then we end up having a session with no PD
associated with that. At the point, we just bail out of the timer
callback as the session is already deleted.