linux_ionic: fix bad stats when device down (#9820)
PS-989
in rhel 7.3 after "ifconfig enp97s0 down", "ethtool -S enp97s0"
shows large stats

PS-990
after interface flap or mtu change, ethtool stats is resetting to 0

In the great TxRx queue replumbing, the tx and rx qcq array creation
was put off from the lif creation to a later time, now at ionic_lif_open(),
and they are now torn down at ionic_lif_close(), so that they can be
rebuilt in a different way (e.g. different q count, MTU change) without
unloading and reloading the driver.  However, when the qcqs are torn down,
the collected sw stats go away with them, so the output from ethtool -S
would all go back to 0 after an MTU change.

This patch arranges for creating the Tx and Rx stats structs at lif
creation time and links them into the qcq when the queues are created.
To do this, we change the txqcqs[] from an array of qcq pointers to
an array of structs that have a qcq pointer and a stats pointer, and
we fill the stats pointer at lif creation time.

Signed-off-by: Shannon Nelson <snelson@pensando.io>