Based on the output files attached to the 2 issues PS-984 and PS-995, the initial observation is that whenever there was a failure, the failure was either on a test in async mode, or the failure was on a test following an async test. Sometimes, one failure might immediately be followed by another failure, but the point still is that an async test was present and seemed to induce the failure. (#9962)
Secondly, both PS-984 and PS-995 have been unreproducible on local runs, whether as part of a series of runs or standalone, suggesting that the issue was neither related to sync mode nor to the checksum operation.

Lastly, while investigating the issue, it was noted that for async cp+pad (when it is last in the chain), the pnso code was programming HW to generate interrupt for *each and every* chain, as opposed to doing it to only the last chain in a batch. For large test repeat count, this could assert million of interrupts. So the first fix attempt will be to address this defect area.