eth_rx: reset dma command stack in sg error case (#10099)
In the sg error case, reset the local DMA command index
instead of saved index, otherwise, the next dma command in this
stage will be written at the top of the dma command stack instead
of the bottom of the stack. Also, the DMA command in the next stage
will write at the bottom of the stack because it uses the saved
index.