Do not invoke Close() on TLS provider on server shutdown (#10201)
GetDefaultTLSProvider(svcName) returns the same provider if svcName
is same, as to minimize the round-trips to the certificates API
when multiple clients or server with the same identity are instantiated.
Since the provider can be shared, it should not be Closed() as part
of client/server cleanup. Client was already NOT cleaning up, server
was actually doing it.

As part of this we also change the identity of the server that was
sharing the TLS provider as it should have its own.