linux_ionic: reduce rx buffer size wrt MTU (#10230)
The Linux driver was allowing packets to pass through that were
slightly larger than the FreeBSD was allowing for the same MTU.
This comes down to the size of the rx buffer that is posted to
the NIC for receive.  The Linux driver was originally set up to
allow space for the optional vlan tag, which gave an additional
4 bytes of packet length when no vlan tag was in use.  Upstream
Linux drivers seem to have several different interpretations of
how this should be configured.  We decided to be consistent
between our own drivers.

PS-1084

Signed-off-by: Shannon Nelson <snelson@pensando.io>