PXE Infra Changes: (#10312)
1. Adding infra for PXE boot which can take testbed.json and get the
rught image loaded on to the system
2. Testing on the infra for performing PXE boot and running classis,
connectivity bundle done for few OSs
Note: Testing pending on this and changes needed as it moves forward.
Lab Tools:
Packaging naples and nic file for console or ssh command runs.
Multiple blocks and changes done as part of this commit. Ned testing to
be done in extention which will follow thru.