TCP Proxy: OOO fixes and window update tx2rx support (#10682)
- Do not ack every OOO packet that is queued back from Tx to Rx. When an
in-order packet is received, we send an ack only when all OOO packets have been
dequeued and queued to SERQ. This is the behavior in Linux (as confirmed by a
pcap where drops are induced)

- TODO: Send an ack when a packet becomes in-order but no packets get dequeued
from OOO queue.

- Fix a bug that caused OOO packets queued back to Rxdma to be launched at
offset 0 (instead of offset 64 that is needed for RxDMA). This caused the wrong
CB address (Tx instead of Rx) to be loaded and was corrupting Tx CB. The fix
for this is to have packets queued to RxDMA to have table0_valid = 1.

- Add support for WIN UPD trigger from tx2rx via a feedback packet. This is
currently used at the end of dequeueing from all OOO queues. This can also be
used when wanting to trigger a window update when the application queue (SERQ)
becomes empty.

- Enhance OOO DOL test case to check for ack at the end of dequeue from OOQ.

- Still keep ooo_queue disabled until more testing is done