This commit fixes one particular cp+pad+chaining issue, where the "chaining" service is one that (#10667)
requires the packed SGL format (for example, full block hash/checksum/decompression).

On post-compression padding, P4+ chainer makes 3 updates to the next-in-chain source SGL: set (partial) last block
length in tuple A, set pad_buffer with pad_len in tuple B, and terminate the last SGL. Note that tuple B could very
well be in the 1st tuple of the last SGL.

The issue is, if the app happens to supply a buffer list such that all tuples in the last SGL get filled by the
sonic driver, there might not be an extra SGL for P4+ chainer to use. For example, the app submits a 32K compression with
a threshold of 12K (minus hdr) and the destination buffer list is also 12K (which becomes the source of the next-in-chain
service). Sonic will provide exactly one packed SGL to P4+ chainer and, if the compression result is > 8K, the chaining
operation will fail because the extra SGL for the pad_buffer/pad_len isn't available.

The fix is to have sonic link in an extra (empty) SGL specifically for the above use case, that is, cp+pad+chaining.
Also note that per the issue description above, crypto as the next-in-chain service would not be affected because
it uses the sparse AOL format and, hence, no fix is required.