linux_ionic: block changes to mnic mtu (#10764)
The mnic device talks directly to the Naples internal mnic,
which is very handy for getting into and tweaking the Naples
configuration with tools like penctl.  However, if the host
user were to mess with the host's mnic mtu, it could cause
surprising and odd failures when subsequently running penctl.

There really is no reason to change the link parameters on
the mnic, and there are other features that really don't make
sense on the mnic (e.g. macvlan offload), so this patch sets
up a separate struct netdev_ops for the mnic to better separate
what the mnic can do versus what the standard ionic port can do.

PS-478

Signed-off-by: Shannon Nelson <snelson@pensando.io>