Fix for PS-1234: bypass-onfail + per-block chksum/hash may cause Naples hang or request timeout. (#10839)
This issue is caused by mismatched src and dst lengths in BOF handling.

BOF works off an array of descriptors and a count. The count tells P4+ how many descriptors
to skip in order to get to the start of the "alternate" descriptors. So far the assumption has
been that the same count also specifies the number of alternate descriptors to submit to HW.
In the failure case, the 2 counts weren't meant to be the same. As an example, say M is the
skip count and N is the submission count. If M < N, the test will fail with a timeout.
If M > N, the test will likely hang Naples. The bug in sonic is that, when M != N,
sonic would give a wrong value of M to P4+. The bug in P4+ is the aforementioned assumption
that M==N, so P4+ would have to modified to allow 2 different values