linux_ionic: implement the tx timeout handler (#10911)
* linux_ionic: implement the tx timeout handler

The tx timeout handler has had a TODO for several months
to get filled in.  We had some instances recently where
a timeout was detected and recovery simply fell on its face.
This should help a little.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: clear interupt credits on queue init

When a queue and its interrupt are setup, be sure to clear
any leftover interrupt credit counts.

Implement the ethtool nway-reset feature, which allows the
user to trigger a TxRx queue reset.  If nothing else, this
helps in testing some odd corner cases.

While we're here, we also centralize the reset code.

Signed-off-by: Shannon Nelson <snelson@pensando.io>