NAPLES admission params checks and update (#11074)
Fixes for https://pensando.atlassian.net/browse/VS-156

NAPLES Spec parameters, such as hostname or IP config, may have changed
while the NIC was not part of the cluster. When NAPLES succesfully gets
admitted, if there is an existing SmartNIC object, we should update the
values to reflect the most current one. However, fields like "Admit"
which can only be changed via Venice, should be unaffected.

With this change we also reject NAPLES admission if the configured
Hostname is already in use by another NAPLES. Result and reason
are returned to NAPLES so that they can be inspected using penctl.
We also create a SmartNIC object in Venice to notify user of the
misconfiguration.