linux_ionic: limit queue flush to size of ring (#11160)
The original developer used a -1 where an unsigned value belongs
to signify cleaning all the queue.  This caused much heartache
and pain when throughput was high and we found ourselves trying
to return interrupt credit values that were way out of line,
especially since it started trying to clean the queue multiple
times in one go.  Trying to return obnoxiously high numbers of
interrupt credits made the NIC unhappy and eventually stopped
the interrupts and then queue processing.  This led to a Tx
Timeout in the OS and further unhappiness in the world.

By limiting the queue flush to the actual number of descriptors
in the queue, we don't pass through the queue more than once
in a napi session.  We can come back later if there are
more to clean, and we keep the individual credit return writes
to reasonable values.

Signed-off-by: Shannon Nelson <snelson@pensando.io>