TCP Proxy: Enable TCP OOO (reassembly) queue (#11276)
- Add queue full check for rx2tx OOO queue

- Do not do window calculation when sending dupack as peer does not count it as
a dupack.

- Fix bugs associated with usage of queues 2, 3 and 4 in Rx. (Only queue 1 was
working correctly and getting DMAed to Tx when becoming in-order). Also
rearranged PHV, so that rx2tx queues 1-4 are contiguous and inside a flit
(otherwise phvwrp does not work).

- When a queue becomes in-order and is queued from tx2rx, more packets may have
arrived at Rx and sitting in the Rx OOO queue. Previously we were not
re-scanning this OOQ, unless we get an in-order packet from the peer. Fixed the
logic so that the last ooo tx2rx packet causes the OOO queue to be rescanned.

- tcp-rx uses a bunch of flags in d using .l (local scope). These flags need to
belong in a byte of their own and cannot share the byte with other data that do
not have local scope.

- When sending dupack we don't want to advance ack sequence number unless all
packets from OOO queue have been dequeued. Introduce another dup_rcv_nxt
variable in rx2tx for this purpose, that can retain the old ack sequence number
even when advancing rcv_nxt.

- Set rcv_mss same as smss for now (otherwise we run into issues testing with
small MSS on iperf)