Fixes PS-1299 (#11272)
Name of file storing events has changed which caused a 404 on penctl
show events. Due to the recent update/changes on events file rotation
mechanism, the filename is no more static, it gets created automatically
based on the timestamp. This change creates a symlink with dynamic
files. so, that penctl can read events as like before.

With this change, penctl will show only recent events.

Signed-off-by: Yuva Shankar <yuva29@users.noreply.github.com>