linux_ionic: set default queue for missing macvlan (#11445)
When supporting multi-lif for macvlan offload, the queue array
can possibly be sparse, so we have to watch out for the stack
selecting a tx queue that doesn't actually exist, typically in
a race when a macvlan channel has just been removed.  If seen,
we just run the packet through queue 0 - there should be few
enough to not make any noticeable impact.  We also add a call
to reset the qdiscs when deleting a macvlan, which should take
care of most of the problem before we get to the Tx.

Signed-off-by: Shannon Nelson <snelson@pensando.io>