fwupdate support for packages containing their own installer. (#11892)
For all fwupdate commands requiring a package file (e.g. install), fwupdate
will look for an "installer" node in the MANIFEST, like this:

    "installer": {
        "name": "foo-installer",
        "verify": {
            "algorithm": "sha512",
            "hash": "05ead9..."
        }
    }

If found, then the new installer file ("foo-installer" in this example) is
extracted from the package file and checksummed.  If it passes checksum
then it is exec'd with an initial argument --custom-installer followed by all
the original fwupdate arguments, and it is the expected to act on the original
request (e.g. install).  The --custom-installer flag avoids recustion.